requirements.txt
requests>=2.28.0
pandas>=1.5.0
python-dotenv>=1.0.0
pathlib>=1.0.1